refactor(browser-repl): remove lib directory
