+++ packages/browser-repl/src/autocompleter/autocompleter.ts
