+++ packages/browser-repl/src/autocompleter/shell-api-autocompleter.ts
