+++ packages/browser-repl/src/electron-runtime/electron-interpreter-environment.ts
