+++ packages/browser-repl/src/iframe-runtime/iframe-interpreter-environment.ts
