+++ packages/browser-repl/src/interpreter/preprocessor/collect-top-level-lexical-context.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '../../../../testing/chai';
+import { expect } from '../../../testing/chai';
 
 import { parse } from '@babel/parser';
