+++ packages/browser-repl/src/interpreter/preprocessor/collect-top-level-lexical-context.ts
