+++ packages/browser-repl/src/interpreter/preprocessor/save-and-restore-lexical-context.ts
