+++ packages/browser-repl/src/interpreter/preprocessor/transform-command-invocation.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '../../../../testing/chai';
+import { expect } from '../../../testing/chai';
 import { transformCommandInvocation } from './transform-command-invocation';
 
 describe('transformCommandInvocation', () => {
