+++ packages/browser-repl/src/interpreter/preprocessor/wrap-object-literal.ts
